+++ include/mruby_integration/core/drawing.hpp
@@ -1,6 +1,0 @@
-#pragma once
-
-#include "mruby.h"
-
-void
-append_core_drawing(mrb_state*);